refactor: move library "tests" into examples
