+++ src/lib.rs
@@ -221,84 +221,3 @@ impl Ambisonic {
             .play(input, BstreamConfig::new().with_position(pos))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::thread::sleep;
-    use std::time::Duration;
-
-    #[test]
-    fn it_works() {
-        let engine = AmbisonicBuilder::new().build();
-
-        let source = rodio::source::SineWave::new(440);
-        let mut first = engine.play_omni(source);
-
-        first.set_position([1.0, 0.0, 0.0]);
-        sleep(Duration::from_millis(1000));
-
-        let source = rodio::source::SineWave::new(330);
-        let mut second = engine.play_omni(source);
-
-        second.set_position([-1.0, 0.0, 0.0]);
-        sleep(Duration::from_millis(1000));
-
-        first.stop();
-        sleep(Duration::from_millis(1000));
-
-        second.set_position([0.0, 1.0, 0.0]);
-        sleep(Duration::from_millis(1000));
-
-        drop(engine);
-
-        sleep(Duration::from_millis(1000));
-    }
-
-    #[test]
-    fn move_sound() {
-        let scene = AmbisonicBuilder::default().build();
-
-        let source = rodio::source::SineWave::new(440);
-        let mut sound = scene.play_at(source, [50.0, 1.0, 0.0]);
-
-        // move sound from right to left
-        sound.set_velocity([-10.0, 0.0, 0.0]);
-        for i in 0..1000 {
-            sound.adjust_position([50.0 - i as f32 / 10.0, 1.0, 0.0]);
-            sleep(Duration::from_millis(10));
-        }
-        sound.set_velocity([0.0, 0.0, 0.0]);
-    }
-
-    #[test]
-    fn bench() {
-        use rodio::Source;
-
-        let scene = AmbisonicBuilder::default().build();
-
-        let mut f: u64 = 1;
-        for _ in 0..850 {
-            f = (f + f * f * 7 + f * f * f * 3 + 1) % 800;
-            let source = rodio::source::SineWave::new(440).amplify(0.001);
-            let _ = scene.play_omni(source);
-        }
-
-        sleep(Duration::from_secs(10));
-    }
-
-    #[test]
-    fn hrir() {
-        let cfg = HrtfConfig::default();
-        let scene = AmbisonicBuilder::default().with_config(cfg.into()).build();
-
-        let source = sources::Noise::new(48000);
-
-        let mut sound = scene.play_at(source, [50.0, 1.0, 0.0]);
-
-        for i in 0..1000 {
-            sound.adjust_position([(500 - i) as f32 / 10.0, 1.0, 0.0]);
-            sleep(Duration::from_millis(10));
-        }
-    }
-}
